12-11 commit
+PostMapping
+GetMapping
+PutMapping
+DeleteMapping
>id converted to long from int

+++ src/main/java/com/csb/samplefullstack/entity/Employee.java
@@ -8,12 +8,12 @@ import javax.persistence.*;
 public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employeeId;
+    private Long employeeId;
     private String name;
     private String location;
     private String department;
 
-    public Employee(int employeeId, String name, String location, String department) {
+    public Employee(Long employeeId, String name, String location, String department) {
         this.employeeId = employeeId;
         this.name = name;
         this.location = location;
@@ -24,11 +24,11 @@ public class Employee {
         
     }
 
-    public int getEmployeeId() {
+    public Long getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(int employeeId) {
+    public void setEmployeeId(Long employeeId) {
         this.employeeId = employeeId;
     }
